refactor(appserver): split resilient adapter
